requirements.txt
APScheduler==3.6.3
autopep8==1.5.4
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==2.2
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
Markdown==3.3.3
oauth2client==4.1.3
oauthlib==3.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
python-decouple==3.4
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
urllib3==1.26.2
whitenoise==5.2.0
